Ionic Modals, Forms
When the user clicks on the "Add Comment" option in the popover, bring
up a modal containing the form that enables the user to add comments on
the dish
The form should contain three fields: Rating, Name and Comment. Use a
Select element for the Rating, and a textarea for Comment fields.
The form should be defined a template named dish-comment.html
The modal is properly configured within DishDetailController
When the user submits the comment, it should be added to the list of
comments for the dish.